// models/interfaces/av_efi_schema.ts
export interface CategorizedThing {
    /** Designates type, e.g. to distinguish different identifiers (GNDResource vs. VIAFResource) */
    category: string,
}


/**
 * Base class defining slots that are common to all levels of the WVMI metadata model
 */
export interface MovingImageRecord extends CategorizedThing {
    /** Also record some metadata about the PID itself rather than the identified object */
    described_by?: DescriptionResource,
    /** Associate event(s) with a moving image record */
    has_event?: Event[],
    /** Record PID in this slot when exporting data from the PID system. Use local identifiers instead when PIDs have not been registered yet. The latter is suitable for transferring data to the agent responsible for registering PIDs */
    has_identifier?: MovingImageResource[],
    /** Indicate a dataset this record has been generated or derived from. For example, a converter generating AVefi moving image records from data in some other schema may record the original identifier here. */
    has_source_key?: string[],
    /** FIAF Moving Image Cataloguing Manual 1.3.5, 2.3.3 */
    in_language?: Language[],
    /** Additional title(s) associated with the work / variant, manifestation, or item. */
    has_alternative_title?: Title[],
    /** Primary title to be displayed in search results etc. The type should be PreferredTitle for works / variants and TitleProper for manifestations / items. If not available, type must be SuppliedDevisedTitle, instead. */
    has_primary_title: Title,
}


/**
 * Metadata about the PID rather than the identified object, i.e. who modified the PID metadata record when, making what changes
 */
export interface DescriptionResource {
    /** Link to revision history of this PID */
    has_history?: string,
    /** Identifier for the responsible party as an URI suitable for linked data */
    has_issuer_id: string,
    /** Name of the responsible party */
    has_issuer_name: string,
    /** Timestamp (in UTC) for the latest modification to any field in the PID metadata record */
    last_modified?: string,
}
// ...
export interface GeographicName extends CategorizedThing {
    /** Alternative human-readable name(s) for a thing. Whereas has_name provides the preferred display name for the described entity, alternatives can be recorded here in order to be indexed in search engines, for instance */
    has_alternate_name?: string[],
    /** Human-readable name for a thing. This is to be treated as the preferred display label in a UI context, whereas has_alternate_name can provide additional terms, e.g. for matching in search operations */
    has_name: string,
    /** See [AuthorityResource doucmentation](AuthorityResource.md) for accepted identifiers */
    same_as?: AuthorityResource[],
}
// ...
export interface Activity extends CategorizedThing {
    /** Agent involved in some activity related to the moving image resource */
    has_agent: Agent[],
    /** See specific class documentation for controlled vocabulary applicable to the type slot, respectively */
    type: string,
}
// ...
export interface Agent extends CategorizedThing {
    /** Alternative human-readable name(s) for a thing. Whereas has_name provides the preferred display name for the described entity, alternatives can be recorded here in order to be indexed in search engines, for instance */
    has_alternate_name?: string[],
    /** For natural persons, always use the convention "family name, given name" */
    has_name: string,
    /** See [AuthorityResource doucmentation](AuthorityResource.md) for accepted identifiers */
    same_as?: AuthorityResource[],
    /** See specific class documentation for controlled vocabulary applicable to the type slot, respectively */
    type: string,
}
// ...
export interface Event extends CategorizedThing {
    /** Associate activity (and subsequently agents) with event */
    has_activity?: Activity[],
    /** Date (or interval/period) when an event has taken place. A subset of ISO 8601 is supported, more specifically, EDTF conformance level 0 as well as qualifiers ? (uncertain date) and ~ (approximate date). See type ISODate definition for details */
    has_date?: string,
    /** Location associated with an event, e.g. the country where the principal offices or production facilities of the production company are located should be associated with the production event */
    located_in?: GeographicName[],
}
// ...
export interface ProductionEvent extends Event {
    /** See specific class documentation for controlled vocabulary applicable to the type slot, respectively */
    type?: string,
}
// ...
export interface Title {
    /** Human-readable name for a thing. This is to be treated as the preferred display label in a UI context, whereas has_alternate_name can provide additional terms, e.g. for matching in search operations */
    has_name: string,
    /** Provide normalised form, e.g. for sorting by title. Only use this slot if value actually if different from has_name */
    has_ordering_name?: string,
    /** FIAF Moving Image Cataloguing Manual A.2 */
    type: string,
}
// ...
export interface Language {
    /** [ISO 639-2 code](https://id.loc.gov/vocabulary/iso639-2.html) for the Representation of Names of Languages (Part 2: Alpha-3) */
    code?: string,
    /** FIAF Moving Image Cataloguing Manual 2.3.3.2 */
    usage: string,
}
// ...
export interface MovingImageRecordContainer {
    /** Root slot holding the moving image metadata record, i.e. metadata describing a work/variant, manifestation or item. See also the Introduction of the FIAF Moving Image Cataloguing Manual */
    has_record: MovingImageRecord,
}


/**
 * Root class for all identifiers from some kind of authority or public register widely accepted in the community
 */
export interface AuthorityResource extends CategorizedThing {
    /** A unique identifier for a thing */
    id: string,
}


/**
 * Either a persistent or local identifier for AVefi compliant moving image records. See subclasses for details
 */
export interface MovingImageResource extends AuthorityResource {
}